backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.6
email-validator==2.1.0
openai==1.10.0
fastapi-mail==1.4.1
python-dotenv>=1.0.0
llama-index-core==0.10.12
llama-index-llms-ollama==0.1.2
llama-index-embeddings-ollama==0.1.2
llama-index-vector-stores-qdrant==0.1.3
qdrant-client==1.7.3
pypdf==4.0.1
# --- Phase 1 additions ---
alembic==1.13.1
pgvector==0.2.5
sentence-transformers==2.7.0
aioredis==2.0.1
slowapi==0.1.9
